fix 同じファイルに差分を反映するように変更
